test: Fix check-realms
On Debian systems /etc/resolv.conf can be a symbolic link.
In order to use 'chattr +i' properly in our test, we just
remove the link and write our own file instead of trying
to overwrite the existing one.

To join a domain on Debian, a few extra packages are needed.
Also, curl shouldn't insist on secure certificates.