ws: Include the RedHat font
Make it the preferred font for Cockpit pages. We can't switch to it
exclusively, as older or newer pages/bridge may run against an
older/newer cockpit-ws (which ships the fonts). For the same reason ws
still needs to ship the Open Sans fonts for a while.

Use it in the HTML guide as well.

Preparatory work for issue #11914