storage: Fix layout of StorageLink:s in content tabs
They always filled the whole row, which was wrong.  (The text was left
aligned, which made it look correct.)  Recently, the text started to
be centered for some reason, which made it look wrong as well.

Wrapping them all in a extra <div> seems to magically fix the layout.
Now the link is only as wide as its content.