test: Fix double destruction in TestMachines.testCreatePXE
THIS IS ALL WRONG, STAY TUNED.

The tests destroys pxe-guest twice, once before and once after
expecting it in the output of "virsh list".  The first destruction is
likely a mistake, since it doesn't match the comment above it, and
we can't expect it to be listed after destruction.

This probably suceeded most of the time because the world is one big
race.  Maybe virt-install didn't get far enough when the first destroy
was run and would restart it later.  It fails reliably in a new
fedora-33 image, however.