storage: Make sure indices are always up-to-date
By always calling update_indices() before triggering a "changed"
event.  Normally, the VDO feature doesn't change anything that
requires re-running of update_indices.  But it can happen that the VDO
feature triggers the very first update and in that case
update_indices() needs to run.

So let's just make a function that does everything needed and use it
consistently.

We also enable the VDO feature last so that the other features can
initialize themselves with less risk of being interrupted by a
"changed" signal.