storage: Also recognize o.fd..Error.ServiceUnknown as "not-found"
The C bridge closes the channel with "not-found" as appropriate, via
watching the name and the connection status.  The P bridge just
returns the D-Bus error from whatever it was doing (probably a
Introspect call for a 'add-watch' request).  The latter seems much
simpler and clients should just handle it.

Or, cockpit.js can stuff `problem: "not-found"` into a ServiceUnknown
error.