containers/ws: Enable beibooting and install standard cockpit pages
This provides similar functionality as the Client, so that the ws
bastion container can connect to machines which don't have any Cockpit
packages installed.

Install cockpit-system and cockpit-networkmanager into the container
(we'll likely expand this in the future, but let's start small). Take
care to only unpack the RPMs, to avoid installing e.g. NetworkManager
and other big dependencies into the container -- they are useless there.

Update the documentation accordingly. While at it, add an introduction
paragraph what Cockpit is.

https://issues.redhat.com/browse/COCKPIT-954