shell: Require confirmation before connecting to remote machines
This requires calling connect_host() in reaction to the "connect"
event of the ShellState so that the warning dialog will be shown
before attempting any connection.

A consequence of that is that now login attempts are always
interactive, and show the password prompts immediately instead of only
showing the "Troubleshoot" curtain with a button to launch those
dialogs.