shell: Only navigate to a host once connected
The old behavior was to navigate always and try to login
afterwards. If the login failed, the "Troubleshoot" curtain would be
shown instead of the pages from the host.

The new behavior will first try to connect to the host and only
navigate to it when the connection has succeeded.